requirements.txt
# EVOLVE-MEM: Production Requirements
# Recommended Python version: >=3.9, <3.12 for best compatibility
# Core dependencies for the three-tier memory architecture

# Core ML and Data Processing
numpy==1.26.4
scikit-learn==1.4.2
scipy==1.11.4
pandas==2.2.2

# Embeddings and Vector Database
sentence-transformers==2.6.1
chromadb==0.4.24

# LLM Integration
google-generativeai==0.8.5
requests==2.31.0
python-dotenv==1.0.1

# Evaluation Metrics
rouge-score==0.1.2
bert-score==0.3.13
nltk==3.8.1

# For METEOR metric, run:
#   python -c "import nltk; nltk.download('wordnet'); nltk.download('omw-1.4')"

# Utilities
# (dateparser is used for advanced date extraction)
dateparser==1.2.0

# Optional dependencies (uncomment if needed):
# hdbscan==0.8.33           # For advanced clustering
# transformers==4.41.2      # For local LLM support
# torch==2.3.0              # For local LLM support
# rich==13.7.1              # For advanced CLI output/logging (not used in main code)
# pydantic==1.10.14         # For data validation (not used in main code)
# openai==1.30.1            # For OpenAI LLM support (not used in current code) 